add extra info card
