+++ app/components/mandatarissen/mandataris-card.hbs
@@ -51,12 +51,6 @@
           {{/if}}
         </div>
       </div>
-      <div class="au-o-grid__item au-u-5-6">
-        <div class="au-c-description-label"><div>Beleidsdomein(en)</div></div>
-        <div class="au-c-description-value">{{await
-            this.formattedBeleidsdomein
-          }}</div>
-      </div>
       {{#if @mandataris.linkToBesluit}}
         <div class="au-o-grid__item au-u-5-6">
           <div class="au-c-description-label"><div>Link naar besluit</div></div>
