+++ app/components/mandatarissen/mandataris-card.js
@@ -12,14 +12,4 @@ export default class MandatarisCardComponent extends Component {
       ? this.args.mandataris.heeftLidmaatschap.get('binnenFractie').get('naam')
       : '';
   }
-
-  get formattedBeleidsdomein() {
-    const beleidsdomeinenPromise = this.args.mandataris.beleidsdomein;
-    if (!beleidsdomeinenPromise.length) {
-      return [];
-    }
-    return beleidsdomeinenPromise.then((beleidsdomeinen) => {
-      return beleidsdomeinen.map((item) => item.label).join(', ');
-    });
-  }
 }
