+++ app/components/mandatarissen/mandataris-extra-info-card.hbs
@@ -1,0 +1,9 @@
+<div class="au-o-box--small au-u-background-gray-100">
+  <div class="au-o-grid au-o-grid--tiny">
+    <div class="au-o-grid__item au-u-5-6">
+      <div class="au-c-description-label"><div>Beleidsdomein(en)</div></div>
+      <div class="au-c-description-value">{{await this.formattedBeleidsdomein}}
+      </div>
+    </div>
+  </div>
+</div>
