+++ app/components/mandatarissen/mandataris-extra-info-card.js
@@ -1,0 +1,13 @@
+import Component from '@glimmer/component';
+
+export default class MandatarisExtraInfoCardComponent extends Component {
+  get formattedBeleidsdomein() {
+    const beleidsdomeinenPromise = this.args.mandataris.beleidsdomein;
+    if (!beleidsdomeinenPromise.length) {
+      return [];
+    }
+    return beleidsdomeinenPromise.then((beleidsdomeinen) => {
+      return beleidsdomeinen.map((item) => item.label).join(', ');
+    });
+  }
+}
