+++ app/templates/mandatarissen/persoon/mandataris.hbs
@@ -27,6 +27,10 @@
   </div>
 
   <div class="au-o-box">
+    <Mandatarissen::MandatarisExtraInfoCard @mandataris={{@model.mandataris}} />
+  </div>
+
+  <div class="au-o-box">
     <Mandatarissen::MandatarisEditPrompt
       class="au-c-collapsible au-u-margin-top-small"
       @mandataris={{@model.mandataris}}
